new service

+++ Coop-Backend/src/main/java/ca/mcgill/ecse321/coop/service/CoopSystemService.java
@@ -4,6 +4,7 @@ import java.sql.Time;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
+import java.util.HashSet;
 
 import org.apache.catalina.User;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,7 +25,10 @@ public class CoopSystemService {
 	EmployerRepository employerRepository; //Create a repository for employers
 	
 	@Autowired
-	DocumentRepository documentRepository; //Create a repository for employers
+	CoopUserRepository coopUserRepository; //Create a repository for employers
+	
+	@Autowired
+	DocumentRepository documentRepository; //Create a repository for documents
 	
 	@Autowired
 	EventNotificationRepository eventNotificationRepository;
@@ -47,6 +51,11 @@ public class CoopSystemService {
 		 
 		CoopSystem coopSystem= new CoopSystem();
 		coopSystem.setId("Main data");
+		coopSystem.setCoopJobs(new HashSet<CoopJob>());
+		coopSystem.setDocuments(new HashSet<Document>());
+		coopSystem.setCoopUsers(new HashSet<CoopUser>());
+		coopSystem.setMessages(new HashSet<Message>());
+		coopSystem.setEventNotifications(new HashSet<EventNotification>());
 		coopSystemRepository.save(coopSystem);
 		return coopSystem;
 	}
@@ -72,7 +81,7 @@ public class CoopSystemService {
 	@Transactional
 	public Student createStudent (String username) //create a student with unique username
 	{	//make sure this username is unique for any employer of student
-		if(studentRepository.existsById(username) || employerRepository.existsById(username) || username==null) //check the username not used before
+		if(username==null || studentRepository.existsById(username) || employerRepository.existsById(username) ||coopUserRepository.existsById(username) ) //check the username not used before
 		{
 			return null;
 		}
@@ -80,8 +89,16 @@ public class CoopSystemService {
 		s.setUsername(username);
 		CoopSystem coopSystem= getCoopSystem();
 		s.setCoopSystem(coopSystem);  //set its top level containing class to the one we are using
+		s.setAuthoredDocuments(new HashSet<Document>());
+		s.setCoopJobs(new HashSet<CoopJob>());
+		s.setPersonalDocuments(new HashSet<Document>());
+		s.setReceivedMessages(new HashSet<Message>());
+		s.setSentMessages(new HashSet<Message>());
 		
 		coopSystem.getCoopUsers().add(s); // add it to the list of users in the system isntance
+		coopSystemRepository.save(coopSystem);
+		studentRepository.save(s);
+		coopUserRepository.save(s);
 		
 		return s;
 	}
@@ -89,7 +106,7 @@ public class CoopSystemService {
 	@Transactional
 	public Employer createEmployer (String username)
 	{   
-		if(studentRepository.existsById(username) || employerRepository.existsById(username) || username==null) //check the username not used before
+		if( username==null || coopUserRepository.existsById(username) || studentRepository.existsById(username) || employerRepository.existsById(username) ) //check the username not used before
 		{
 			return null;
 		}
@@ -97,31 +114,76 @@ public class CoopSystemService {
 		e.setUsername(username);
 		CoopSystem coopSystem= getCoopSystem();
 		e.setCoopSystem(coopSystem);
+		e.setAuthoredDocuments(new HashSet<Document>());
+		e.setCoopJobs(new HashSet<CoopJob>());
+		e.setArchivedInterns(new HashSet<Student>());
+		e.setReceivedMessages(new HashSet<Message>());
+		e.setSentMessages(new HashSet<Message>());
 		
 		coopSystem.getCoopUsers().add(e);
 		
+		coopSystemRepository.save(coopSystem);
+		employerRepository.save(e);
+		coopUserRepository.save(e);
 		return e;
 	}
 	
 	@Transactional
-	public void setPassword(CoopUser a, String password) //set the password for a CoopUser
+	public void setPassword(String username, String password) //set the password for a CoopUser
 	{
-		if(a!=null && password!=null) {a.setPassword(password);}
+		if(username==null || password==null) {return;}
+		CoopUser s=getCoopUser(username);
+		if(s==null) {return;}
+		s.setPassword(password);
+		saveCoopUser(s);
 	}
 	
 	@Transactional
-	public void setStudentPermissions(Student s, boolean allowCV, boolean allowTranscript)
+	public CoopUser getCoopUser(String username) {
+		if(username==null) {return null;}
+		if(studentRepository.existsById(username)) {return studentRepository.findById(username).get();}
+		else if(employerRepository.existsById(username)) {return employerRepository.findById(username).get();}
+		return null;
+	}
+	
+	@Transactional
+	public Student getStudent(String username) {
+		if(username==null) {return null;}
+		if(studentRepository.existsById(username)) {return studentRepository.findById(username).get();}
+		return null;
+	}
+	
+	@Transactional
+	public Employer getEmployer(String username) {
+		if(username==null) {return null;}
+		if(employerRepository.existsById(username)) {return employerRepository.findById(username).get();}
+		return null;
+	}
+	
+	@Transactional
+	public void saveCoopUser (CoopUser s) {
+		if(s==null) {return;}
+		if(s instanceof Student) {coopUserRepository.save(s);studentRepository.save((Student) s);}
+		else {coopUserRepository.save(s);employerRepository.save((Employer) s);}
+	}
+	
+	@Transactional
+	public void setStudentPermissions(String username, boolean allowCV, boolean allowTranscript)
 	{
+		if(username==null) {return;}
+		Student s= getStudent(username);
 		if(s!=null)
 		{
 			if(allowCV || !allowTranscript)
 			{
-				s.setAllowCV(allowCV); s.setAllowTranscript(allowTranscript); return;
+				s.setAllowCV(allowCV); s.setAllowTranscript(allowTranscript);
 			}
 			else { //if the parameters allow Transcript but forbid cv, change nothing
 				
 			}
 		}
+		coopUserRepository.save(s);
+		studentRepository.save(s);
 	}
 	
 	
@@ -203,53 +265,63 @@ public class CoopSystemService {
 	
 	
 	@Transactional
-	public void deleteStudent(Student a)
+	public void deleteStudent(String username)
 	{
-		if(a!=null) {studentRepository.delete(a);}
+		Student a= getStudent(username);
+		if(a!=null) {studentRepository.delete(a);}//coopUserRepository.delete(a);
 	}
 	
 	@Transactional
-	public void deleteEmployer(Employer a)
+	public void deleteEmployer(String username)
 	{
-		if(a!=null) {employerRepository.delete(a);}
+		Employer a =getEmployer(username);
+		if(a!=null) {employerRepository.delete(a);}//coopUserRepository.delete(a);
 	}
 	
 	@Transactional
-	public void deleteCoopUser(CoopUser a) // delete any type of user
+	public void deleteCoopUser(String username) // delete any type of user
 	{
+		CoopUser a=getCoopUser(username);
 		if(a!=null)
 		{
 			if(a instanceof Student) //if this is a student
 			{
-				studentRepository.delete((Student)a);
+				studentRepository.delete((Student)a);//coopUserRepository.delete(a);
 			}
 			else // if this is an employer
 			{
-				employerRepository.delete((Employer) a);
+				employerRepository.delete((Employer) a);//coopUserRepository.delete(a);
 			}
 		}
 	}
 	
 	@Transactional
-	public void setPersonalDocuments(Student stu, ArrayList<Document> docs)
+	public void setPersonalDocuments(String username, ArrayList<String> docIds)
 	{
-		if(stu!=null && docs!=null && stu.getPersonalDocuments().size()< 3) //make sure they are 2 max	
+		if(username==null) {return;}
+		Student stu =getStudent(username);
+		if(stu!=null && docIds!=null ) //make sure they are 2 max	
 		{
-			for(Document doc: docs)
+			stu.setPersonalDocuments(new HashSet<Document>());
+			for(String id: docIds)
 			{
-				stu.getPersonalDocuments().add(doc);
-				doc.setStudent(stu);
+				Document doc=findDocumentByDocumentId(id);
+				if(doc!=null) {stu.getPersonalDocuments().add(doc);}
+				
 			}
 		}
+		coopUserRepository.save(stu);
+		studentRepository.save(stu);
 	}
 	
 	@Transactional
-	public Document createDocument(String documentId, CoopUser author)
+	public Document createDocument(String documentId, String authorUsername, DocumentType type)
 	{	//unique id
-		if(documentId==null || documentRepository.existsById(documentId) || author ==null)
+		if(documentId==null || documentRepository.existsById(documentId) || authorUsername ==null)
 		{
 			return null;
 		}
+		CoopUser author= getCoopUser(authorUsername);
 		CoopSystem c= getCoopSystem();
 		Document d= new Document(); //create a new document instance
 		
@@ -258,12 +330,15 @@ public class CoopSystemService {
 		d.setCoopSystem(c); // set its top-level class to the coop system we are currently using
 		d.setSubmissionDate(new Date(System.currentTimeMillis())); //set submission date and time to current time
 		d.setSubmissionTime(new Time(System.currentTimeMillis()));
+		d.setType(type);
 		
 		
 		author.getAuthoredDocuments().add(d); //add this document to the list this user authored
 		c.getDocuments().add(d); // add this document to the list of documents in the top level class
 		
-		
+		coopSystemRepository.save(c);
+		documentRepository.save(d);
+		saveCoopUser(author);
 		
 		return d;
 	}
@@ -280,9 +355,10 @@ public class CoopSystemService {
 	}
 	
 	@Transactional
-	public ArrayList<Document> findDocumentsByAuthor(CoopUser author)
+	public ArrayList<Document> findDocumentsByAuthor(String authorUsername)
 	{
-		if(author==null) {return null;}
+		if(authorUsername==null) {return null;}
+		CoopUser author= getCoopUser(authorUsername);
 		Set<Document> set= author.getAuthoredDocuments();
 		ArrayList<Document> list= new ArrayList<Document>();
 		for(Document d: set) {
@@ -293,8 +369,10 @@ public class CoopSystemService {
 	
 	
 	@Transactional
-	public void deleteDocument(Document d)
+	public void deleteDocument(String id)
 	{
+		if(id==null) {return;}
+		Document d= findDocumentByDocumentId(id);
 		if(d!=null) {documentRepository.delete(d);}
 	}
 	
@@ -321,15 +399,19 @@ public class CoopSystemService {
 		for(Employer em: this.getAllEmployers())
 		{
 			em.getEventNotifications().add(e);  //add this event notification to the list of notifications for all employers
+			employerRepository.save(em);
 		}
 		
-		
+		coopSystemRepository.save(s);
+		eventNotificationRepository.save(e);
 		return e;
 	}
 	
 	@Transactional
-	public void setEventNotificationSettings(EventNotification e, Event typeEvent, String location, Date date, Time startTime, Time endTime)
+	public void setEventNotificationSettings(String eventName, Event typeEvent, String location, Date date, Time startTime, Time endTime)
 	{
+		if(eventName==null) {return;}
+		EventNotification e = findEventNotificationByName(eventName);
 		if(e==null) {return;}
 		if(endTime.compareTo(startTime)>=0 && date.compareTo(new Date(System.currentTimeMillis()))>=0) //change only if endTime>=startTime and the date is after the day we are setting this
 		{
@@ -339,6 +421,7 @@ public class CoopSystemService {
 			e.setStartTime(startTime);
 			e.setType(typeEvent);
 		}
+		eventNotificationRepository.save(e);
 	}
 	
 	
@@ -366,18 +449,21 @@ public class CoopSystemService {
 	}
 	
 	@Transactional
-	public void deleteEventNotification(EventNotification e)
+	public void deleteEventNotification(String eventname)
 	{
+		EventNotification e = findEventNotificationByName(eventname);
 		if (e!=null) {eventNotificationRepository.delete(e);}
 	}
 	
 	@Transactional
-	public Message createMessage(String messageId, CoopUser sender, CoopUser receiver, String content, ArrayList<Document> attachements)
+	public Message createMessage(String messageId, String senderName, String receiverName, String content, ArrayList<String> attachementNames)
 	{	//unique id, sender and receiver exist
-		if(messageId==null || sender== null ||  receiver==null || messageRepository.existsById(messageId))
+		if(messageId==null || senderName== null ||  receiverName==null || messageRepository.existsById(messageId))
 		{
 			return null;
 		}
+		CoopUser sender=getCoopUser(senderName);
+		CoopUser receiver= getCoopUser(receiverName);
 		
 		Message m= new Message();
 		m.setMessageId(messageId);
@@ -386,18 +472,23 @@ public class CoopSystemService {
 		m.setCoopSystem(getCoopSystem());
 		m.setDate(new Date(System.currentTimeMillis()));
 		m.setTime(new Time(System.currentTimeMillis()));
+		m.setAttachements(new HashSet<Document>());
 		if(content!=null) {m.setContent(content);}
-		if(attachements!=null)
+		if(attachementNames!=null)
 		{
-			for(Document doc: attachements)
+			for(String docId: attachementNames)
 			{
-				m.getAttachements().add(doc);
+				Document doc= findDocumentByDocumentId(docId);
+				if(doc!=null) {m.getAttachements().add(doc);}
 			}
 		}
-		
-		getCoopSystem().getMessages().add(m);
+		CoopSystem c=getCoopSystem();
+		c.getMessages().add(m);
 		sender.getSentMessages().add(m); //add this message to the list of messages for each user
 		receiver.getReceivedMessages().add(m);
+		coopSystemRepository.save(c);
+		messageRepository.save(m);
+		saveCoopUser(sender);saveCoopUser(receiver);
 		return m;
 	}
 	
@@ -410,8 +501,10 @@ public class CoopSystemService {
 	}
 	
 	@Transactional
-	public ArrayList<Message> findMessagesBySender(CoopUser sender)
+	public ArrayList<Message> findMessagesBySender(String senderName)
 	{
+		if(senderName==null) {return null;}
+		CoopUser sender =getCoopUser(senderName);
 		if(sender==null) {return null;}
 		ArrayList<Message> list= new ArrayList<Message>();
 		Set<Message> set= sender.getSentMessages();
@@ -423,9 +516,10 @@ public class CoopSystemService {
 	}
 	
 	@Transactional
-	public ArrayList<Message> findMessagesByReceiver(CoopUser receiver)
+	public ArrayList<Message> findMessagesByReceiver(String receiverName)
 	{
-		if(receiver==null) {return null;}
+		if(receiverName==null) {return null;}
+		CoopUser receiver=getCoopUser(receiverName);
 		ArrayList<Message> list= new ArrayList<Message>();
 		Set<Message> set= receiver.getReceivedMessages();
 		for(Message m: set)
@@ -436,10 +530,13 @@ public class CoopSystemService {
 	}
 	
 	@Transactional
-	public ArrayList<Message> findMessagesBySenderAndReceiver(CoopUser sender, CoopUser receiver)
+	public ArrayList<Message> findMessagesBySenderAndReceiver(String senderName, String receiverName)
 	{
+		if(receiverName==null || senderName==null) {return null;}
+		CoopUser sender=getCoopUser(senderName);
+		CoopUser receiver=getCoopUser(receiverName);
 		if(receiver==null || sender==null) {return null;}
-		ArrayList<Message> listo= findMessagesBySender(sender);
+		ArrayList<Message> listo= findMessagesBySender(senderName);
 		ArrayList<Message> listn = new ArrayList<Message>();
 		for(Message m: listo) //get the messages where threciever is indeed the receiver
 		{
@@ -452,31 +549,42 @@ public class CoopSystemService {
 	}
 	
 	@Transactional
-	public void deleteMessage(Message m)
+	public void deleteMessage(String messageId)
 	{
+		if(messageId==null) {return;}
+		Message m= findMessageByMessageId(messageId);
 		if(m!=null) {messageRepository.delete(m);}
 	}
 	
 	@Transactional
-	public CoopJob createCoopJob(String jobId, Employer employer, Student student)
+	public CoopJob createCoopJob(String jobId, String employerName, String studentName)
 	{
-		if(jobId==null|| employer==null || student==null || coopJobRepository.existsById(jobId) ) {
+		if(jobId==null|| employerName==null || studentName==null || coopJobRepository.existsById(jobId) ) {
 		return null;
 		}
+		Student student=getStudent(studentName);
+		Employer employer=getEmployer(employerName);
+		CoopSystem c=getCoopSystem();
+		if(student==null || employer==null){return null;}
 		CoopJob job =new CoopJob();
 		job.setJobId(jobId);
 		job.setEmployer(employer);
 		job.setIntern(student);
-		job.setCoopSystem(getCoopSystem());
+		job.setCoopSystem(c);
+		job.setCoopJobDocuments(new HashSet<Document>());
 		
 		employer.getCoopJobs().add(job);
 		student.getCoopJobs().add(job);
-		getCoopSystem().getCoopJobs().add(job);
+		c.getCoopJobs().add(job);
 		if(!employer.getArchivedInterns().contains(student)){ // if this the first time this student does a coop
 															// a coop with this employer, add this employer to
 															//archived interns
 			employer.getArchivedInterns().add(student);
 		}
+		coopJobRepository.save(job);
+		coopSystemRepository.save(c);
+		studentRepository.save(student);coopUserRepository.save(student);
+		employerRepository.save(employer);coopUserRepository.save(employer);
 		return job;
 	}
 	
@@ -489,8 +597,10 @@ public class CoopSystemService {
 	}
 	
 	@Transactional
-	public ArrayList<CoopJob> findCoopJobsByEmployer(Employer em)
+	public ArrayList<CoopJob> findCoopJobsByEmployer(String employerName)
 	{
+		if(employerName==null) {return null;}
+		Employer em=getEmployer(employerName);
 		if(em==null) {return null;}
 		Set<CoopJob> set= em.getCoopJobs();
 		ArrayList<CoopJob> list= new ArrayList<CoopJob>();
@@ -502,8 +612,10 @@ public class CoopSystemService {
 	}
 	
 	@Transactional
-	public ArrayList<CoopJob> findCoopJobsByStudent(Student s)
+	public ArrayList<CoopJob> findCoopJobsByStudent(String studentName)
 	{
+		if(studentName==null) {return null;}
+		Student s=getStudent(studentName);
 		if(s==null) {return null;}
 		Set<CoopJob> set= s.getCoopJobs();
 		ArrayList<CoopJob> list= new ArrayList<CoopJob>();
@@ -515,9 +627,12 @@ public class CoopSystemService {
 	}
 	
 	@Transactional
-	public ArrayList<CoopJob> findCoopJobsByEmployerAndStudent(Employer em, Student s)
+	public ArrayList<CoopJob> findCoopJobsByEmployerAndStudent(String employerName, String studentName)
 	{
-		if(s==null|| em==null) {return null;}
+		if(studentName==null|| employerName==null) {return null;}
+		Student s=getStudent(studentName);
+		Employer em=getEmployer(employerName);
+		if(s==null || em==null) {return null;}
 		Set<CoopJob> set= s.getCoopJobs();
 		ArrayList<CoopJob> list= new ArrayList<CoopJob>();
 		for (CoopJob job: set)
@@ -528,12 +643,14 @@ public class CoopSystemService {
 	}
 	
 	@Transactional
-	public void deleteCoopJob(CoopJob job) {
+	public void deleteCoopJob(String jobId) {
+		CoopJob job=findCoopJobByJobId(jobId);
 		if(job!=null) {coopJobRepository.delete(job);}
 	}
 	
 	@Transactional
-	public void setCoopJobSettings(CoopJob job, Date startDate, Date endDate, String name, CoopState state) {
+	public void setCoopJobSettings(String jobId, Date startDate, Date endDate, String name, CoopState state) {
+		CoopJob job=findCoopJobByJobId(jobId);
 		if(job==null || endDate.compareTo(startDate)<=0) { //chnage nothing if job is null, or enddate<=startDate
 			return;
 		}
@@ -542,14 +659,18 @@ public class CoopSystemService {
 		
 		if(name!=null) {job.setName(name);}
 		if(state!=null) {job.setState(state);}
+		coopJobRepository.save(job);
 		
 	}
 	
 	@Transactional
-	public void addDocumentToCoopJob(CoopJob job, Document d)
+	public void addDocumentToCoopJob(String jobId, String documentId)
 	{
+		CoopJob job=findCoopJobByJobId(jobId);
+		Document d=findDocumentByDocumentId(documentId);
 		if(job==null || d==null) {return;}
 		job.getCoopJobDocuments().add(d);
+		coopJobRepository.save(job);
 	}
 	
 	@Transactional
